+++ sqry-lsp/CHANGELOG.md
@@ -6,7 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [10.0.2](https://github.com/verivus-oss/sqry/compare/v10.0.1...v10.0.2) - 2026-04-27
+## [10.0.4](https://github.com/verivus-oss/sqry/compare/v10.0.3...v10.0.4) - 2026-04-28
+
+### Other
+- update Cargo.lock dependencies
+## [10.0.3](https://github.com/verivus-oss/sqry/compare/v10.0.2...v10.0.3) - 2026-04-28
+
+### Fixed
+- *(vscode)* restore workspace status and provenance## [10.0.2](https://github.com/verivus-oss/sqry/compare/v10.0.1...v10.0.2) - 2026-04-27
 
 ### Other
 - update Cargo.lock dependencies